# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EarthSync Server API
  description: API for managing users, key exchange, data ingest, and historical data retrieval (spectrograms and peaks). Includes enhanced server-side peak detection and basic transient detection. Archives older data to database. Returns structured, timestamped spectrogram history.
  version: 1.1.14 # Version bump reflects history API change
servers:
  - url: http://localhost:3000 # Modified to remove backticks for compatibility
    description: Development Server (HTTP)
paths:
...
  /history/hours/{hours}:
    get:
      summary: Retrieve historical spectrogram data (structured, timestamped)
      tags: [Data]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: hours
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 168 } # Updated max hours
          description: Number of hours of historical data to retrieve (1-168).
        - name: detectorId
          in: query
          required: false
          schema: { type: string, example: 'detector1', maxLength: 50 }
          description: Optional. Filter data by a specific detector ID.
      responses:
        '200': { $ref: '#/components/responses/HistorySpectrogramSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /history/range:
    get:
      summary: Retrieve historical spectrogram data by time range (structured, timestamped)
      tags: [Data]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: startTime
          in: query
          required: true
          schema: { type: string, format: date-time }
          description: Start timestamp (ISO 8601 format).
        - name: endTime
          in: query
          required: true
          schema: { type: string, format: date-time }
          description: End timestamp (ISO 8601 format).
        - name: detectorId
          in: query
          required: false
          schema: { type: string, example: 'detector1', maxLength: 50 }
          description: Optional. Filter data by a specific detector ID.
      responses:
        '200': { $ref: '#/components/responses/HistorySpectrogramSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /history/peaks/hours/{hours}:
    get:
      summary: Retrieve historical detected peak data
      tags: [Data]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: hours
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 168 } # Updated max hours
          description: Number of hours of historical peak data to retrieve (1-168).
        - name: detectorId
          in: query
          required: false
          schema: { type: string, example: 'detector1', maxLength: 50 }
          description: Optional. Filter peak data by a specific detector ID.
      responses:
        '200': { $ref: '#/components/responses/HistoryPeaksSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /history/peaks/range:
    get:
      summary: Retrieve historical detected peak data by time range
      tags: [Data]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: startTime
          in: query
          required: true
          schema: { type: string, format: date-time }
          description: Start timestamp (ISO 8601 format).
        - name: endTime
          in: query
          required: true
          schema: { type: string, format: date-time }
          description: End timestamp (ISO 8601 format).
        - name: detectorId
          in: query
          required: false
          schema: { type: string, example: 'detector1', maxLength: 50 }
          description: Optional. Filter peak data by a specific detector ID.
      responses:
        '200': { $ref: '#/components/responses/HistoryPeaksSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        lat: { type: number, format: float, example: 40.7128, minimum: -90, maximum: 90 }
        lon: { type: number, format: float, example: -74.0060, minimum: -180, maximum: 180 }
      required: [lat, lon]
...
    DetectedPeak:
      type: object
      properties:
        freq:
          {
            type: number,
            format: float,
            description: 'Detected peak frequency in Hz',
            example: 7.81,
          }
        amp: { type: number, format: float, description: 'Detected peak amplitude', example: 15.3 }
        qFactor:
          {
            type: number,
            format: float,
            nullable: true,
            description: 'Estimated Quality Factor (Center Freq / FWHM)',
            example: 4.5,
          }
        trackStatus:
          {
            type: string,
            enum: [new, continuing],
            description: 'Indicates if the peak is new or part of an ongoing track',
            example: 'continuing',
          }
        trackId: {
            type: string,
            format: uuid,
            nullable: true,
            description: 'Unique identifier for the peak track',
            example: 'f47ac10b-58cc-4372-a567-0e02b2c3d479',
          } # Added trackId
      required: [freq, amp] # qFactor, trackStatus, trackId might not always be present initially
    TransientInfo:
      type: object
      description: 'Information about detected transient events.'
      properties:
        type:
          {
            type: string,
            enum: [none, broadband, narrowband, error],
            description: 'Type of transient detected.',
            example: 'narrowband',
          }
        details:
          {
            type: string,
            nullable: true,
            description: 'Details about the detected transient.',
            example: 'Narrowband signal detected near 4.0 Hz (Amp: 12.1, Delta: 11.1)',
          }
      required: [type]
    HistoricalPeakEntry:
      type: object
      properties:
        ts:
          {
            type: number,
            format: int64,
            description: 'Timestamp of the peak detection (Unix Milliseconds)',
            example: 1711994676123,
          }
        peaks:
          {
            type: array,
            description: 'Array of peaks detected at this timestamp.',
            items: { $ref: '#/components/schemas/DetectedPeak' },
          }
      required: [ts, peaks]
    DetectorPeakHistory:
      type: object
      properties:
        detectorId: { type: string, example: 'detector1' }
        peaks:
          {
            type: array,
            description: 'Array of peak detection events, sorted by timestamp.',
            items: { $ref: '#/components/schemas/HistoricalPeakEntry' },
          }
      required: [detectorId, peaks]
    SpectrogramDataPoint:
      type: object
      description: 'A single timestamped spectrogram data point with associated transient info.'
      properties:
        ts:
          {
            type: number,
            format: int64,
            description: 'Timestamp (Unix milliseconds)',
            example: 1711994676123,
          }
        spectrogram: {
            type: array,
            items: { type: number, format: float },
            description: 'Array of downsampled amplitude values for this timestamp.',
            # Example length would be RAW_FREQUENCY_POINTS / DOWNSAMPLE_FACTOR
          }
        transientInfo: { $ref: '#/components/schemas/TransientInfo' }
      required: [ts, spectrogram, transientInfo]
    DetectorSpectrogramHistoryStructured: # NEW Schema for structured response
      type: object
      description: 'Structured historical spectrogram data for a single detector.'
      properties:
        detectorId: { type: string, example: 'detector1' }
        location: { $ref: '#/components/schemas/Location' }
        dataPoints:
          {
            type: array,
            description: 'Array of timestamped spectrogram data points, sorted chronologically.',
            items: { $ref: '#/components/schemas/SpectrogramDataPoint' },
          }
      required: [detectorId, location, dataPoints]
...
    ErrorResponse:
      type: object
      properties:
        error:
          {
            type: string,
            description: 'A description of the error.',
            example: 'Invalid username or password',
          }
      required: [error]
  responses:
...
    HistorySpectrogramSuccess: # UPDATED Response schema
      description: Historical spectrogram data retrieved successfully (structured).
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/DetectorSpectrogramHistoryStructured' }
    HistoryPeaksSuccess:
      description: Historical peak data retrieved successfully.
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/DetectorPeakHistory' }
...
    BadRequest:
      description: Invalid input provided (e.g., failed validation, incorrect format).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: Authentication failed (e.g., missing/invalid/expired JWT token).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    TooManyRequests:
      description: Rate limit exceeded for the requested endpoint.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    InternalServerError:
      description: An unexpected error occurred on the server.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from the /login endpoint. Include in the Authorization header as 'Bearer {token}'."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API Key required only for the /data-ingest endpoint.'
...